requirements.txt
aio-pika==9.4.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
coverage==7.10.7
cycler==0.12.1
DateTime==5.5
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.117.1
flake8==7.3.0
fonttools==4.60.1
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.9
matplotlib==3.10.6
mccabe==0.7.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
passlib[bcrypt]==1.7.4
pillow==11.3.0
pluggy==1.6.0
psycopg[binary]>=3.1
pycodestyle==2.14.0
pydantic==2.11.9
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
PyJWT==2.10.1
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pwdlib[argon2]
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0
zope.interface==8.0.1
